Move AnalysisDir, AnalysisProject & AnalysisSample into new module 'analysis'.
