+++ auto_process_ngs/test/qc/test_utils.py
@@ -9,7 +9,7 @@ import os
 from bcftbx.JobRunner import SimpleJobRunner
 from auto_process_ngs.mock import MockAnalysisProject
 from auto_process_ngs.mock import UpdateAnalysisProject
-from auto_process_ngs.utils import AnalysisProject
+from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.qc.utils import verify_qc
 from auto_process_ngs.qc.utils import report_qc
